fish_bell option to enable or disable bell on reader_t::flash().
fish_bell's behaviour is documented in doc_src/language.rst.

The option was added because visual (and especially audio) feedback after
failed user interaction is usually unnecessary and can be extremely annoying.
It is only useful over bad networks, when it isn't clear whether the
interaction failed or there is network lag.